Refactor(Files): Addressing was wrong for building
